# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Memtara API
  version: 0.1.0
  summary: Zero-knowledge attribute disclosure, and the signed attestations that carry its verdicts off-platform.
  description: |
    Memtara lets a consumer prove a *predicate* about themselves — income above a
    threshold, residency valid, no sanctions match — without disclosing the
    attribute behind it. Proofs are generated on the holder's device against one
    of four Noir circuits; this API authenticates users, syncs the encrypted
    vault, orchestrates the disclosure-request lifecycle, verifies submitted
    proofs against the real compiled circuits via Barretenberg, and issues
    short-lived Ed25519-signed attestations that any JOSE-speaking relying party
    can validate offline.

    **This API never generates proofs.** Generating one requires the plaintext
    attribute as a witness, so a server that could prove would be a server that
    holds plaintext. See `ARCHITECTURE.md`. Every "verified" verdict here comes
    from running the real circuit's verifier over bytes a client produced.

    **Compliance context.** `docs/REGULATORY_MATRIX.md` maps this system, clause
    by clause, against the CBUAE *Guidance Note on Consumer Protection and the
    Responsible Adoption and Use of AI and ML by LFIs* (`CBUAE_EN_6958_VER1`).
    Predicate responses carry the clause ids each disclosure is evidence for.
  license:
    name: See repository
  contact:
    name: Memtara
...
servers:
...
security:
  - sessionToken: []
paths:
...
  /api/v1/products:
    post:
      tags: [Product registry]
      summary: Register a structured product and its suitability terms
      security:
        - orgApiKey: []
      description: |
        Scoped to the calling org. Two banks may register the same ISIN
        independently, on different terms, and neither can see the other's —
        the key is `(org_id, product_isin)`, not the bare ISIN, because a
        competitor's minimum-income threshold is a commercially sensitive fact.

        `approved_by_risk_committee` defaults to **false**, which is the only
        safe default for a flag whose purpose is to gate assessment.

        Re-registering an existing ISIN is a `409`, not a silent overwrite:
        rewriting the terms every in-flight assessment quotes is far more likely
        to be a mistake than an intent, and `PATCH` exists for the intent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Malformed ISIN, empty name, or a term outside its bounds.
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          description: Already registered for this organisation.
          $ref: "#/components/responses/Error"
    get:
      tags: [Product registry]
      summary: List this organisation's products
      security:
        - orgApiKey: []
      parameters:
        - name: approved_only
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: |
            What an advisor-facing product picker should ask for, so an
            unapproved instrument never appears as recommendable.
      responses:
        "200":
          description: The catalogue.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/products/{isin}:
    parameters:
      - name: isin
        in: path
        required: true
        schema: { type: string, example: XS1234567890 }
    get:
      tags: [Product registry]
      summary: Read one product
      security:
        - orgApiKey: []
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: |
            Not in *this* organisation's registry. Another tenant's product is
            indistinguishable from one that does not exist.
          $ref: "#/components/responses/Error"
    patch:
      tags: [Product registry]
      summary: Amend terms or record risk-committee approval
      security:
        - orgApiKey: []
      description: |
        Every field is optional; only what is supplied changes. The audit event
        records the before and after of each changed field, because "who lowered
        the minimum income, and when" is the first question anyone reviewing a
        mis-sold product asks.

        An amendment applies to assessments opened **after** it. Assessments
        already open keep the terms they were opened with.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmendProduct"
      responses:
        "200":
          description: The amended product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: A term outside its bounds.
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        An opaque, revocable session token from any auth flow. Opaque rather
        than a JWT precisely so it can be killed in one query — the product's
        revocation guarantees depend on it.
    orgApiKey:
      type: http
      scheme: bearer
      description: |
        A relying party's API key, presented in the same `Authorization: Bearer`
        header as a session token. Routes readable by "the org that created it
        or the user it is for" try the presented credential as both.
...
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    Product:
      type: object
      required: [id, org_id, product_isin, product_name, risk_level, min_income,
                 min_liquidity, max_concentration_percent, approved_by_risk_committee,
                 check_digit_valid, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        product_isin: { type: string, example: XS1234567890 }
        product_name: { type: string }
        risk_level: { type: integer, minimum: 1, maximum: 5 }
        min_income: { type: integer, format: int64 }
        min_liquidity: { type: integer, format: int64 }
        max_concentration_percent: { type: integer, minimum: 0, maximum: 100 }
        approved_by_risk_committee:
          type: boolean
          description: |
            Not a label. `issue-wealth-request` returns 409 while this is false.
        check_digit_valid:
          type: boolean
          description: |
            The ISO 6166 check digit result, reported rather than enforced. The
            illustrative identifiers used throughout this project fail it, so
            enforcing would reject the cases the feature was specified against.
            A deployment fed by a real product master should treat `false` here
            as a data-quality alarm.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateProduct:
      type: object
      required: [product_isin, product_name, risk_level, min_income, min_liquidity,
                 max_concentration_percent]
      properties:
        product_isin: { type: string }
        product_name: { type: string }
        risk_level: { type: integer, minimum: 1, maximum: 5 }
        min_income: { type: integer, format: int64, minimum: 0 }
        min_liquidity: { type: integer, format: int64, minimum: 0 }
        max_concentration_percent: { type: integer, minimum: 0, maximum: 100 }
        approved_by_risk_committee: { type: boolean, default: false }
    AmendProduct:
      type: object
      description: Every field optional; only what is supplied changes.
      properties:
        product_name: { type: string }
        risk_level: { type: integer, minimum: 1, maximum: 5 }
        min_income: { type: integer, format: int64, minimum: 0 }
        min_liquidity: { type: integer, format: int64, minimum: 0 }
        max_concentration_percent: { type: integer, minimum: 0, maximum: 100 }
        approved_by_risk_committee: { type: boolean }
...